requirements.txt
cryptography==48.0.0
fastapi==0.136.1
PyJWT==2.8.0
python-multipart==0.0.27
requests==2.33.1
uvicorn==0.46.0